requirements.txt
wheel
selenium
python-telegram-bot
flask
pyngrok
requests